requirements.txt
requests
python-dotenv
python-slugify
transliterate
hypothesis
pytest
